// TwitterWEB/src/models/UserProfileCardModel.ts
export class UserProfileCardModel {
  id: string;
  createdDate: Date;
  fullname: string;
  username: string;
  personalInfo: string;
  location: string;
  personalWebSiteURL: string;
  profilePicPath: string;
  backgroundPath: string;
  followerCounter: number;
  followingCounter: number;
  followFlag: boolean;
}